refactor(decomp): move storage engine-port traits to proximadb-storage-traits crate (link 7)
Capstone of the root-crate decomposition: hoists ~3.3K lines of the
src/storage/traits/ module (the UnifiedStorageFormat engine-port-traits
contract) to a new crates/storage/proximadb-storage-traits/ crate. All
root-deps were pre-hoisted by the 10-PR port-inversion chain
(#896/#901/#902/#906/#910/#913/#929/#930/#933).

Crate contents (moved + rerouted):
- lib.rs (UnifiedStorageFormat core trait + UnifiedMetricsCollector +
  CollectionStats + SegmentIngestResult) — all crate:: refs rerouted to
  proximadb_{proto,kernel,search-types,storage-common,storage-ports,tenant}
- query.rs (StorageQueryContext, RlsRecordPredicate, ParsedQuantizationConfig)
- types.rs (FlushParameters, CompactionParameters, StorageEngineStrategy re-export)
- results.rs (FlushResult, CompactionResult, EngineStatistics, EngineHealth)

Trait split (Approach C — gaps 5+6):
- get_filesystem_factory + 4 staging methods STAY ROOT on a new
  EngineFilesystemAccess extension trait (they reference the root-internal
  FilesystemFactory type — cannot move to the crate).
- The impl_engine_identity! macro STAYS ROOT (::...FilesystemFactory
  with #[macro_export] resolves to the defining crate).
- The 7 ISP traits (StorageCompactor/Identity/Reader/Writer/Scan/Lifecycle/
  Metrics) stay ROOT — they reference root-internal trait_components.
- The capabilities() method STAYS ROOT as a free function
  engine_capabilities() — it returns CapabilitySet from the query-contract
  layer, which storage cannot depend upward into (CI layering guard).
- document.rs + observability.rs STAY ROOT — their traits
  (DocumentStorageOperations, ObservabilityStorageOperations,
  MultiModelStorage) are implemented for foreign types
  (ObservabilityService from the observability-engine crate) which would
  violate Rust orphan rules if the traits were foreign.

All 9 engines (SST/VIPER/HELIX/NOVA/SWIFT/RAPTOR/SEQUOIA/CEDAR/TST) +
10 mock/test impls updated: get_filesystem_factory moved from
impl UnifiedStorageFormat to impl EngineFilesystemAccess.

This unblocks link 8 — the engines extraction (the 188K-symbol jobs=2 +
open-core-contract payoff).
