+++ crates/storage/proximadb-storage-traits/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+name = "proximadb-storage-traits"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Core storage-engine port traits (UnifiedStorageFormat + parameter/result types) for ProximaDB — the contract between the service layer and the pluggable storage engines, hoisted out of the root crate as the capstone of the root-crate decomposition (link 7)"
+
+[lib]
+name = "proximadb_storage_traits"
+path = "src/lib.rs"
+
+[dependencies]
+anyhow.workspace = true
+async-trait.workspace = true
+chrono.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+# Foundation proto + leaf types
+proximadb-proto.workspace = true
+proximadb-records.workspace = true
+proximadb-kernel.workspace = true          # CollectionIdentity + CompactBatchId (BatchId alias)
+proximadb-tenant.workspace = true          # UnifiedUserContext + RbacTenantContext (RBAC)
+proximadb-search-types.workspace = true    # SearchParams (UnifiedSearchParams) + BlockPruneMode + OptimizedSearchRecord
+proximadb-distance-kernel.workspace = true # DistanceMetric
+proximadb-quantization-types.workspace = true # QuantizationType
+proximadb-data-model.workspace = true      # DataModel (MultiModelStorage)
+# Storage-tier
+proximadb-storage-common.workspace = true  # StorageEngineType
+proximadb-storage-ports.workspace = true   # StorageEngineStrategy + ScanStrategy/ScanIterator + CapabilityFactory + path_resolver
+
+[lints]
+workspace = true
