+++ src/storage/engines/cedar/mod.rs
@@ -396,23 +396,6 @@ impl UnifiedStorageFormat for CedarEngine {
         StorageFormatStrategy::Cedar
     }
 
-    fn get_filesystem_factory(&self) -> &FilesystemFactory {
-        use std::sync::OnceLock;
-        static FACTORY: OnceLock<FilesystemFactory> = OnceLock::new();
-        FACTORY.get_or_init(|| {
-            futures::executor::block_on(async {
-                FilesystemFactory::create(FilesystemConfig::default())
-                    .await
-                    .unwrap_or_else(|_| {
-                        #[allow(clippy::panic)]
-                        {
-                            panic!("Failed to create filesystem factory for CEDAR engine")
-                        }
-                    })
-            })
-        })
-    }
-
     async fn collect_engine_metrics(&self) -> Result<HashMap<String, serde_json::Value>> {
         DocumentStorageEngine::collect_metrics(self).await
     }
@@ -439,6 +422,25 @@ impl UnifiedStorageFormat for CedarEngine {
 
     async fn do_compact(&self, _params: &CompactionParameters) -> Result<CompactionResult> {
         Ok(CompactionResult::default())
+    }
+}
+
+impl crate::storage::traits::EngineFilesystemAccess for CedarEngine {
+    fn get_filesystem_factory(&self) -> &FilesystemFactory {
+        use std::sync::OnceLock;
+        static FACTORY: OnceLock<FilesystemFactory> = OnceLock::new();
+        FACTORY.get_or_init(|| {
+            futures::executor::block_on(async {
+                FilesystemFactory::create(FilesystemConfig::default())
+                    .await
+                    .unwrap_or_else(|_| {
+                        #[allow(clippy::panic)]
+                        {
+                            panic!("Failed to create filesystem factory for CEDAR engine")
+                        }
+                    })
+            })
+        })
     }
 }
