+++ tests/sks_integration_test.rs
@@ -114,30 +114,6 @@ mod sks_integration_tests {
             ) -> Result<Vec<proximadb::core::search::results::OptimizedSearchRecord>> {
                 Ok(vec![])
             }
-
-            fn get_filesystem_factory(
-                &self,
-            ) -> &proximadb::storage::persistence::filesystem::FilesystemFactory {
-                // For testing, create a minimal filesystem factory
-                use proximadb::storage::persistence::filesystem::FilesystemFactory;
-                use std::sync::OnceLock;
-                static FACTORY: OnceLock<FilesystemFactory> = OnceLock::new();
-                FACTORY.get_or_init(|| {
-                    // For testing, create a basic filesystem factory synchronously
-                    // In real usage, this would be created with proper async initialization
-                    match tokio::runtime::Handle::try_current() {
-                        Ok(handle) => {
-                            handle.block_on(async {
-                                FilesystemFactory::create(proximadb::storage::persistence::filesystem::FilesystemConfig::default()).await.unwrap()
-                            })
-                        }
-                        Err(_) => {
-                            // If no async runtime, create minimal factory for testing
-                            panic!("FilesystemFactory requires async runtime for initialization")
-                        }
-                    }
-                })
-            }
         }
 
         Arc::new(MockStorageEngine)
